Always expect events to be contiguous.
Previously it was assumed that the earliest known event was a good starting point, but in reality analysis never has a complete view of the world to be able to make any decisions about the data it is receiving. New data could easily arrive that invalidates any assumptions. Now nothing is assumed. Proceed only when one particular event is encountered. Note that is not quite complete - it doesn't cover holes that can occur early in the data stream.

#rb jb
#jira ue-127376
#rnx
#preflight 615aec4ce3b24d0001496bb3

[CL 17706661 by Martin Ridgers in ue5-main branch]