Tasks: removed invalid `verify` as it's triggered in an valid case:
Thread1 can call TaskEvent.Trigger(), which unlocks the task, and before it calls verify(TryExecute()), Thread2 starts TaskEvent retraction (because Thread2 started waiting on its subsequent task), and first it gets execution permission. then it proceeds to complete retraction as the task is unlocked but not executed yet. Now Thread1 tries to get execution permissions and fails, but can proceed as normal. its job is done

#fyi devin.doucette
#robomerge 5.0

[CL 17770492 by Andriy Tylychko in ue5-main branch]