Horde: Provide index hint to find jobs query
Adding an index requires modifying the database which isn't allowed (yet) in the canary env. So the index is reference by name instead, which of course isn't great. Might be possible to look up these during the index creation and reference them by a C# instance/type. But considering the perf impact this has, I think it's worth the trade-off.

[CL 18222276 by carl bystrom in ue5-main branch]