Refactored blob storage in Horde.Storage to allow for central handling of all blob store requests.
This lets us add verification of the blob identifier before submitting content to the blob store.
I also added a blob index, a database tracking of which blobs exists, and a optional toggle to use this instead of checking S3 during existence checks (which we think will be faster). Could also support batched requests but that is not yet implemented.
Also promoted the current site setting into the common (JupiterSettings) settings, as we now also use this for the blob index, which we intend to use for in line replication of missing blobs in the future.

Removed the hirearchal blob stores and moved this into the blob service (the new central place to handle all blob store requests). Change this to process blob stores in parallel which will be a small speedup as we start the s3 requests earlier.
Changed GetOldObjects in the IBlobStore to just fetch AllObjects (as we used it for both of those purposes) and now we include the last modified time with the object for the filtering cases.

Added IBufferedPayload for reusable control of how we buffer the payloads we need to handle. This means we will now handle large (>2GB) payloads better in more places. Also added a setting to control when we start buffering to disk (defaults to 2GB payloads, e.g. payloads to large to fit into memory). Could be used to reduce the amount of memory required by Horde.Storage at the cost of requiring more temp drive space, and likely some performance as we then end up writing to disk. This buffering is required as we need to hash all objects before we store them, requiring us to have the entire payload before it can be sent anywhere.

[CL 18308858 by Joakim Lindqvist in ue5-main branch]