Skip an ensure if RenderDoc has been detected.
Static initialisation will happen on a thread other than the game thread due to the way RenderDoc injects its DLL (details in the change's comment). Thus the ensure that validates that the game thread and static-init thread are one and the same always fails. Because te GameThread task tag scope is so early (before engine pre-init), and because there is a lot of machinery being a failed ensure (e.g. crash reporter client, dumping of memory stats), this failing ensure is very fragile and can often go undiscovered. So to put it to rest, avoid the ensure altogether if RenderDoc's around.

Note that the segfaulting ensure() is already fixed in 18280863.

#jira UE-99840
#preflight 61a8813fe01f3610f52be80f

[CL 18350568 by Martin Ridgers in ue5-main branch]