Horde: Add index hint and read consistency parameters to LeaseCollection
A new method override for FindLeasesAsync is also provided with hand-picked index and read consistency. Should result in more performant lease lookups in the autoscale service (which are currently super slow and I/O demanding due to a bad index). A database index class was also factored out to avoid duplicating code.

[CL 18382178 by carl bystrom in ue5-main branch]