Horde Storage
Implemented orphan blob gc based on the blob index. This will check every blob in the blob store for a corresponding blob index, if it doesnt exist we delete the blob. If it does exist we check the object references it has, if none of them exist anymore we delete the blob.

[CL 18395079 by Joakim Lindqvist in ue5-main branch]