Pass the render targets information explicitly to FMeshDrawCommand::GetPipelineStateSortingKey, instead of relying on whatever is currently bound.
This method can be called before the render pass is executed, so the currently bound state is irrelevant. The calling code must extract the correct render target configuration from the pass parameters and send it. Relying on ApplyCachedRenderTargets could lead to incorrect sort keys or even crashes when the method was called in between passes, and the previously bound texture pointers were stale.

#jira none
#rnx
#rb Christopher.Waters
#preflight 61b3363313028c27d2928d57

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 18430152 in //UE5/Release-5.0/... via CL 18435290
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Staging -> Release-Engine-Test) (v897-18405271)

[CL 18435514 by mihnea balta in ue5-release-engine-test branch]