Horde Storage
Removed transaction to update last access time, can result in empty rows in the objects table which will generate an error. If it does we can update the code to just consider this as a non existent object. This will generate much less load on the database, and doesnt seem worth for something which we could just handle.

Also removed other call to UpdateIfAsync for finalizing a blob, which isnt needed as its not actually using any conditional statament (so UpdateAsync is the same thing)

[CL 18458689 by Joakim Lindqvist in ue5-main branch]