Horde: Include hourly rate for using an agent in returned agent and job documents.
Global config file now includes a list of conditions which are evaluated against an agent's properties. The first condition that matches specifies the rate for the agent. This should give us plenty of flexibility in pricing out esoteric configurations dependent on factors like instance type, OS, etc...

Rates are cached in Redis and re-evaluated every 5 minutes.

#fyi Josh.Engebretson
#preflight none

[CL 18469963 by Ben Marsh in ue5-main branch]