Fix BSP undo/redo crash
UModelSceneProxy shares data between Game/Render thread. UModel is typically updated by deregistering its components and flushing render commands before updating the vertex buffers. This happens in PreEditUndo. The resources are then initialized and component reregistered in PostEditUndo.

The previous change to immediately process the list of added primitives after the create action of an actor (AActor::PostEditUndo) breaks this contract since the order of the AActor::PostEditUndo in the transaction list is not guaranteed. This change addresses the issue by skipping the immediate UpdateAllPrimitiveSceneInfo for Brush actors. This is a least invasive solution that restores the prior behavior solely for BSP.

#rb mihnea.balta michael.balzer
#rnx
#jira UE-132580 UE-109385
#preflight 61e0b4c5250b9537f786a97f

#ROBOMERGE-AUTHOR: lonnie.li
#ROBOMERGE-SOURCE: CL 18619153 in //UE5/Release-5.0/... via CL 18619303 via CL 18619522
#ROBOMERGE-BOT: UE5 (Release-Engine-Test -> Main) (v899-18417669)

[CL 18619897 by lonnie li in ue5-main branch]