1) Remote Control - Right-click Rebind is available outside of selected row
Rebind is available. It's a bit confusing to have this context menu not confined to the associated property row.

#jira UE-137305
#rb simon.therriault
#rb jeremie.roy
#preflight 18595842

2) Remote API Panel Shortcut

Being outside Edit Mode does not lock the individual exposed properties as it did in 4.27. Currently, it only seems to affect group headings. For example, properties can still be removed, renamed, and sorted while outside of edit mode in 5.0.

#jira UE-114549
#rb simon.therriault
#rb jeremie.roy
#preflight 18595842

#ROBOMERGE-AUTHOR: paul.vincent
#ROBOMERGE-SOURCE: CL 18626591 in //UE5/Release-5.0/... via CL 18626592 via CL 18626593
#ROBOMERGE-BOT: UE5 (Release-Engine-Test -> Main) (v899-18417669)

[CL 18626595 by paul vincent in ue5-main branch]