FrameRate: add support for parsing FFrameRates from strings that end with drop frame tokens "ndf" or "df"
This allows parsing a frame rate from an FPS string that denotes timecode drop frame-ness with
either "ndf" for non-drop frame or "df" for drop frame as a suffix. Drop frame timecode is only
supported for NTSC_30 (29.97fps) and NTSC_60 (59.94fps) frame rates, so the only FPS strings
that include the "df" token that will parse successfully are "29.97df" and "59.94df". Any other FPS
value that includes the "df" token will result in an error. All NTSC FPS values (23.976, 29.97, and 59.94)
as well as any integer FPS value can include the "ndf" token.

Note however that the presence or absence of the "ndf" and "df" tokens in the string will have
no effect on the resulting FFrameRate. This simply ensures that strings that include those tokens
can be parsed into the correct frame rate.

#rb alejandro.arango
#preflight 61d8e466af44b4362826d48e

#ROBOMERGE-AUTHOR: matt.johnson
#ROBOMERGE-SOURCE: CL 18676930 in //UE5/Release-5.0/... via CL 18676952 via CL 18676965
#ROBOMERGE-BOT: UE5 (Release-Engine-Test -> Main) (v902-18672795)

[CL 18680186 by matt johnson in ue5-main branch]