POST data size was not specified correctly.
Unknown POST sizes cause later libcurl versions to add an Transfer-Encoding header which is not compatible with the Content-Length header. An exert from the libcurl docs; "without chunked transfer, you must specify the size in the request. (Since 7.66.0, libcurl will automatically use chunked encoding for POSTs if the size is unknown.) When providing data with a callback, you must transmit it using chunked transfer-encoding or you must set the size of the data with the CURLOPT_POSTFIELDSIZE"

#rb sbo
#rnx
#preflight 61f255fffc74f46b56569b47

[CL 18751364 by Martin Ridgers in ue5-main branch]