ShaderConductor: Parse ovr_multiview_view_count option
spirv-cross requires the ovr_multiview_view_count option set when using multiview. We pass in the target option in CompileToGlslWithShaderConductor, but we didn't process it in ShaderConductor.

I also had to modify some functionality in the HCT.cmake module that was dependent on local git configuration to determine CRLF vs LF EOL. DXC is originally a git project, and it depends on the local
git config to make decisions. But our setup is not dependent on git, so there can be unintended side-effects depending on the sytem or global git config.

#rb steve.smith, jeannoe.morissette
#jira UE-136124
#preflight 61f9d6681d7ca8ed2d6236a6

#ROBOMERGE-AUTHOR: robert.srinivasiah
#ROBOMERGE-SOURCE: CL 18819448 in //UE5/Release-5.0/... via CL 18819460 via CL 18822892
#ROBOMERGE-BOT: UE5 (Release-Engine-Test -> Main) (v910-18824042)

[CL 18824745 by robert srinivasiah in ue5-main branch]