Fixed incorrect GetCookedSandboxDir() path on Mac.
FPath::ProjectSavedDir() can be influenced by command line arguments such as `-userdir=` or the installed status of the engine and thus can return directories that are not where cooked output is expected to be. The `-installed` command line argument is added to UE_GAME-type targets on Mac, such that ProjectSavedDir() does not return the same path the cooker used (which uses ProjectDir() + "/Saved"). As a result, Mac does not work the same way other desktop platforms work during development.

#rb cm
#rnx
#jira UE-133202
#preflight 61fa83274404d5fade1a6389

[CL 18825484 by Martin Ridgers in ue5-main branch]