DirectLink in Editor - Fixing async task never stopped and preventing UE from closing.
Auto-reconnection task will re-enqueue another reconnection task after running if there are still DirectLink sources to reconnect to. Normally this task would be stopped on module shutdown, but it turns out that the presence of running background tasks prevents Unreal from shutting down, and thus the module was never shutdown. To fix this, we now listen for the on FCoreDelegate::OnPreExit event, and stop any running reconnection task.

#jira UE-141512
#rb JeanLuc.Corenthin
#preflight 61fdd88e163df1051e5ef9b1
#lockdown JeanMichel.Dignard

#ROBOMERGE-AUTHOR: benoit.deschenes
#ROBOMERGE-SOURCE: CL 18889073 in //UE5/Release-5.0/... via CL 18889084 via CL 18889334
#ROBOMERGE-BOT: UE5 (Release-Engine-Test -> Main) (v910-18824042)

[CL 18889346 by benoit deschenes in ue5-main branch]