USD: build using CMake directly on Windows and add usdAbc plugin
This adds a batch script on Windows with the recipe for building USD and
its supporting files using CMake directly. This ensures that USD builds
specifically against the engine's versions of USD's library dependencies
rather than having USD's build script download and build its own versions.
Some of the manual steps involved in preparing a build of USD for installation
in the engine were rolled into this batch script. When the script completes,
the build products are mostly ready to be copied into place, although
plugInfo.json files will still need their LibraryPath to be adjusted.

With that change in place, we can also now build the usdAbc plugin that
is included with USD which adds a file format plugin for reading Alembic
files via USD. Alembic makes use of the half-precision floating type defined
in the Imath library, and USD has not yet been officially updated to adapt
to the separation of Imath from OpenEXR, but a small patch to the USD
CMake build makes this work.

These changes include the updates to the build process as well as a full
rebuild of USD for Windows built using this new setup. This pattern will be
propagated to Linux and Mac once all of the dependencies (particularly Boost)
are in place on those platforms.

#rb daniel.coelho, anousack.kitisa
#preflight 6204480ebf7362cd77faf506

[CL 18946802 by matt johnson in ue5-main branch]