USDImporter: updates to adapt to UsdLuxLightAPI and Ar 2.0 changes in USD 21.11 and later
This is prep work in advance of upgrading the engine's version of USD
from 21.08 to the yet-to-be-released 22.03.

In USD 21.11, UsdLuxLight was removed and replaced with UsdLuxLightAPI,
which is an API schema instead of a typed one.

USD 21.11 also saw the default version of Ar move to 2.0 instead of 1.0.

The changes here should allow the USDImporter plugin for UE to be compiled with
versions both before and after the UsdLux transition, and with either version
of Ar active. However, assets or tests that make use of lights may still need
updating. Those updates will be made in follow-up changes.

#rb daniel.coelho, anousack.kitisa
#preflight 620ac1b4674d5ad0da8f9ec6

[CL 18987326 by matt johnson in ue5-main branch]