[Backout] - CL18952424
This is causing an ensure to hit in Slate applications.

[FYI] Patrick.Boutot, Matt.Kuhlenschmidt
Original CL Desc
-----------------------------------------------------------------
Slate: Clear the uncached draw elements once they are batches and clear the batches elements once they are rendered. We do this to remove hard references to the Glyph Sequence from the previous frame. Cached draw elements are still cached. An old Glyph Sequence may still be a reference in a cached element but it is the responsibility of the SWidget to do the proper invalidation.
#rb matt.kuhlenschmidt
#preflight 62066e4554003c49ad37c3e6

#ROBOMERGE-AUTHOR: vincent.gauthier
#ROBOMERGE-SOURCE: CL 18997554 via CL 18997606 via CL 18998070 via CL 18998217 via CL 18998286 via CL 19002213 via CL 19004425
#ROBOMERGE-BOT: UE5 (Release-Engine-Test -> Main) (v917-18934589)

[CL 19004631 by vincent gauthier in ue5-main branch]