Reset active rota item count returning to do another pass over it.
Some rota items cannot make progress because their next event type is not known yet. They get moved to the end of the rota and the active count is decremented. There are two cases where the rota might be reprocessed from the beginning again; when an initial serial is set, and when the rota shows it has the next event. It is possible that on a subsequent pass through the rota, items that could not make progress previously might be able to this time. However, because the active count was left over from before they were excluded and OnData() would not do all the work it could possibly do.

All this only applies to old P2 traces.

#rb jb
#rnx
#preflight 620cd9a36202a22eeddb71ff

[CL 19014434 by Martin Ridgers in ue5-main branch]