MaterialX: add build script, build products, and module rules for using MaterialX in the engine on Windows
This change includes MaterialX version 1.38.1 built for use in UE5 on Windows.

The module rules do not currently account for installing the MaterialX standard data
libraries (definitions of pattern and shading nodes) anywhere, so prospective clients
of MaterialX may need to reach directly into the MaterialX deployment for now. This
will be addressed in a subsequent change.

Note also that the "cmake" directory of the installation is excluded from the deployed
build products. There are fixes coming in a future release of MaterialX that place the
"cmake" directory correctly inside the "lib" directory, which is necessary to ensure
that the CMake files generated for the various platforms do not collide. When we
upgrade to that future release, those files can then be added per-platform.

#jira UE-142380
#rb danny.kabrane
#preflight 620c2f12565043444caf23b1

[CL 19039130 by matt johnson in ue5-main branch]