Chaos Cloth - Fix collisions against convexes now that they're floats.
Additionally add lots of compile time verification that the ispc structs and enums match the c structs and enums for collision objects. This required adding a new accessor (which I called ISPCDataVerifier) to many of the implicit objects which can report the offset and size of its members.

Highlighted code changes:

TaperedCylinder.h
Switched this to hold FPlaneConcrete instead of FPlane. The difference is that FPlane is an FImplicitObject and otherwise has a bunch of passthroughs to FPlaneConcrete. We never use the fact that these are FImplicitObjects, so we don't need this additional 32 bytes per plane.

PerParticlePBDCollisionConstraint.ispc
On top of updating the convex code to be the correct float types, I also added th e FTaperedCapsule and FTaperedCylinder ispc structs. This makes it easier to verify this data.

ImplicitObject.h, ChaosClothingSimulationCollider.cpp
Technically all TSpheres think they have ImplicitObjectType::Sphere, but all of our code assumes that ImplicitObjectType::Sphere corresponds with TSphere<FReal,3> when casting to the derived type. I added a new typedef for FSphere which is this one type.

#rb michael.lentine
#preflight 620e9200e55d9645848ecf90
#jira UE-143152
#lockdown michael.lentine


#ROBOMERGE-AUTHOR: alex.mcadams
#ROBOMERGE-SOURCE: CL 19098395 via CL 19099712 via CL 19100195 via CL 19100716 via CL 19109898
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v921-19075845)

[CL 19147223 by alex mcadams in ue5-main branch]