ARCore - Use HLSLcc for external texture support
DXC does not support external textures in 5.0. ARCore requires external textures for support. We need to rollback to HLSLcc for the templates.
I did try using ModifyCompilationEnvironment, but the DXC flags aren't set by the time it is invoked.

This is TEMPORARY. Once external texture support lands in our DXC/ShaderConductor backend, we can revert this change to use DXC for ARCore.

#jira UE-126772
#rb Rolando.Caloca
#preflight 621fd1cf37049be51732f6a7

#ROBOMERGE-AUTHOR: robert.srinivasiah
#ROBOMERGE-SOURCE: CL 19227615 in //UE5/Release-5.0/... via CL 19229035
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v921-19075845)

[CL 19231558 by robert srinivasiah in ue5-main branch]