Improve path tracer invalidation through transform changes to reduce the number of spurious invalidations
Rather than performing the change detection on the game thread, perform it only for the rendering primitive that might impact the path tracer. In particular, this avoids updates to the static meshes representing cameras. The camera transforms are handled on a per view basis, which means cameras not being moved will not have their output invalidated.

#jira UE-143951
#preflight 62200f6c943787f1d21ae666
#rb Alejandro.Arango,Simon.Therriault
#lockdown Juan.Canada

#ROBOMERGE-AUTHOR: chris.kulla
#ROBOMERGE-SOURCE: CL 19245026 in //UE5/Release-5.0/... via CL 19249057
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v924-19243027)

[CL 19262938 by chris kulla in ue5-main branch]