Potential Security Issue, Login Screen
Have addeda a FHttpRequestHandler as a RequestPreprocessor which checks whether or not an Requests Header has a corresponding ApiKey to allow the execution, using MD5 hashing to obscure it a bit.
This affects Http Calls and Websocket Calls. Changes Wrapped RCWebsocketRequest to allow holding hashed ApiKeys from the outside. Mainly hooked as an additional Handler in WebRemoteControl.

#jira UE-139210
#rb simon.therriault, amir.keren
#preflight 622113f2c7262e0e538118cf
#lockdown alejandro.arango

#ROBOMERGE-OWNER: simon.therriault
#ROBOMERGE-AUTHOR: patrick.gross
#ROBOMERGE-SOURCE: CL 19252881 in //UE5/Release-5.0/... via CL 19257182
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v924-19243027)

[CL 19263496 by simon therriault in ue5-main branch]