Fix occasional D3D12 crash after allocating a resource via RHICreateTexture2DFromResource and releasing it.
FD3D12DynamicRHI::CreateTextureFromResource() shouldn't call Location.AsStandAlone(), since that adds the allocation to FD3D12Adapter::TrackedAllocationData. However, since the type is then changed to eAliased, the resource will not be removed from TrackedAllocationData when it's released (FD3D12ResourceLocation::ReleaseResource() only calls UpdateStandAloneStats() for eStandAlone resources, as it should). This skews the stats, but more importantly causes an assert if the memory for the released FD3D12ResourceLocation object is reused by a subsequent allocation.

#jira UE-145332
#rnx
#rb Kenzo.TerElst
#ushell-cherrypick of 17608117 by mihnea.balta
#preflight 622b2a8cdf3960e506c60b87
#lockdown Michal.Valient

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 19353645 in //UE5/Release-5.0/... via CL 19356073
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v926-19321884)

[CL 19357789 by mihnea balta in ue5-main branch]