Fix groom asset DDC cache issue.
The HairDescription*Group* struct wasn't updated when the HairDescription struct was updated. This caused the groom data being built & cached upon the 'old' HairDescription data. This CL ensure the HairDescriptionGroup struct is updated at the same time as the HairDescription.

This CL also rename HairDescription and HairDescriptionGroup with a 'Cached' prefix to highlight the fact these data structure are transient.
This CL also change groom asset DDC key so that previous invalid cached data are wiped out.

This issue is critical for MH & has been reported by several users.

#rb kevin.ortegren
#jira MH-5452
#preflight 622b60ee46679d56c3156bf7
#lockdown michal.valient

#ROBOMERGE-AUTHOR: charles.derousiers
#ROBOMERGE-SOURCE: CL 19354477 in //UE5/Release-5.0/... via CL 19356144
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v926-19321884)

[CL 19357816 by charles derousiers in ue5-main branch]