HLSL Syntax Hightlighter: added more type keywords. Merged with the old list of keywords and sorted based on string comparison in descending order. The keywords are from CrossCompiler::EHlslToken.
I feel like exposing an API to query those keywords on CrossCompiler might not be worth the effort. We will need to introduce an extra storage there to stores the types as they are added to their internal map -> filter out type keywords based on a range in the enum class ->  merge it with existing list -> sort the list -> remove duplicates. Could get difficult to debug if one of the steps goes wrong.

#jira UE-144112
#rb halfdan.ingvarsson
#preflight skip

[CL 19376316 by jack cai in ue5-main branch]