DirectLink in Editor - Fixing deadlock caused by recursive read-lock.
When a write-lock is requested, additional read-locks will queue up behind the pending write-lock, even if the FRWLock is currently locked in a read (shared) state. This can cause deadlock when we attempt to read-lock the FRWLock recurisivelly.

In this particular case a deadlock could occur between FDirectLinkManager::GetUriFromSourceHandle() and FDirectLinkManager::OnStateChanged(). To fix this now free the read-lock in FDirectLinkManager::GetOrCreateExternalSource() before calling GetUriFromSourceHandle() to avoid locking recursively.

#jira UE-145959
#rb JeanLuc.Corenthin
#preflight 6233b6798a298a014fb49fa6

[CL 19429241 by benoit deschenes in ue5-main branch]