Packaging: Fixed incorrect generation of layer information for subchunks leading to package generator crashing with an exception.
Subchunks are generated when having MaxChunkSize parameter set to a non-zero value in the project packaging properties. The problem was caused by the generated pakchunklayers.txt having information only for the main chunks, but not for subchunks. The package generator assumes that pakchunklayers.txt has the same number of lines as the generated pakchunklist.txt (storing a list of all chunks and subchunks).

This issue is a regression introduced in CL 7637286, which modified the code to output layer information for empty chunks.

#jira UE-146062
#rb dave.barrett
#preflight 6234adc40820efd0946b55eb

[CL 19437131 by Wojciech Krywult in ue5-main branch]