Fix FLinearColor::ToFColorSRGB handling of large A, Quantize handling of large R/G/B/A.
For large enough floats x (e.g. 1e+7f), x*255 is too large to convert to int. On x86 this turns into INT_MIN which then gets clamped to 0, on ARM it gets saturated and turns into INT_MAX which gets clamped to 255.

Fix it so that over-large values result in 255 (which is the more natural behavior) everywhere and note the conversion rules used for both the sRGB and the linear channels while I'm here. Also update ImageCore optimized kernels to include the QuantizeRound fix.

Confirmed to produce bitwise identical results for textures on AncientGame (i.e. nop on typical content).

#jira  UE-146294
#preflight 6234ffcc48746817f14025dc
#rb martins.mozeiko

[CL 19442181 by fabian giesen in ue5-main branch]