Concert: delay sending Sequencer state sync events to the client until its workspace is synced and finalized
This addresses an issue particularly with nDisplay nodes where the client can receive
the Sequencer state sync event before it has finished processing transactions from
the activity stream. If a sequence open on another client was created or saved by a
transaction in that stream, the nDisplay node may not be ready to open or sync a
Sequencer/player for it until its workspace is finalized.

In addition, a Sequencer/player is opened for any sequence not already open on the
client when the Sequencer state sync event is received. Since the client may have
been launched some time after other clients, it would not have received Sequencer
open events that occurred before it was launched.

This also includes some minor cleanup to event handling and delegate registration and
unregistration, as well as a few small corrections to some docstrings.

#jira UE-144128
#rb jason.walter
#preflight 623511e448746817f140bec1

[CL 19454582 by matt johnson in ue5-main branch]