USDImporter: register translators for boundable and nonboundable lights in USD 21.11 and later
This is another bit of prep work in advance of upgrading the engine's version of USD
from 21.08 to 22.03.

In USD 21.11, UsdLuxLight was removed and replaced with UsdLuxLightAPI,
which is an API schema instead of a typed one. The concrete base schemas for
the built-in light types are now UsdLuxBoundableLightBase and
UsdLuxNonboundableLightBase. When building against a version of USD that
uses the UsdLuxLightAPI and these new(-ish) base schemas, we register
translators for them instead of the now defunct UsdLuxLight schema.

#jira UE-143295
#rb anousack.kitisa, daniel.coelho
#preflight 623b8db188538cd45e1fd22a

[CL 19498866 by matt johnson in ue5-main branch]