Config (ini): Fixed extraction of values from quoted strings, which was off by one.
For instance,

@PrimaryAssetTypesToScan=PrimaryAssetType
+PrimaryAssetTypesToScan=(PrimaryAssetType="Map",AssetBaseClass=/Script/Engine.World,bHasBlueprintClasses=False,bIsEditorOnly=True,Directories=((Path="/Game/Maps")),SpecificAssets=,Rules=(Priority=-1,ChunkId=-1,bApplyRecursively=True,CookRule=Unknown))

when processing the second line PrimaryAssetType key would be extracted as "Ma. It worked correctly for non-quoted values.

NOTE: Based on my review, there were no known consequences of this bug, because these extracted values are exclusively used for comparisons (with all compared values being extracted in the same way). It could cause problems if we had values used as keys differing just by the last character.

#rb Josh.Adams
#preflight 62432ed3f4217035fa6f23aa

[CL 19544197 by Wojciech Krywult in ue5-main branch]