Add Enhanced Input developer and Platform settings.
Create a "UEnhancedInputPlatformData" base class that can be subclassed in a blueprint and placed on specific platforms. By default, this object has a map of Input Contexts that will "redirect" one context to another. This makes it possible to swap out Input Mapping Contexts on specific platforms, with an asset that will simply be added to the input ini of the game.

#jira UE-144850
#preflight 6245cae4c51908059abd88b4
#rb andrew.davidson
#rb dave.jones2

[CL 19575308 by ben hoffman in ue5-main branch]