SourceControl: Prevent unnecessary asset saves.
When choosing Check-In on a directory in the Content Browser, PromptForCheckoutAndSave is called with a list of all loaded packages that reside in that directory hierarchy by FAssetFolderContextMenu::ExecuteSCCCheckIn with bCheckDirty=true and bPromptToSave=true. However, if none of the given packages is dirty, the end result would be that all packages would be saved instead of none of the packages being saved. This is unfortunate as source control systems will -unnecessarily- end up considering those assets as changed.

#jira none
#review
#preflight 624b16ddf4217035fab6fdfb

[CL 19621627 by wouter burgers in ue5-main branch]