UBT: recognize shared libraries with version number suffixes as allowed duplicate build products
Shared libraries on Linux commonly use a scheme of version number suffixes
and/or symlinks to help ensure that client applications can link against the
appropriate version of a library (e.g. "libfoo.so.1.70.0"). This change updates
IsDuplicateBuildProduct() with an additional case to cover these libraries so
that they are not flagged as an error by the build system if multiple steps
produce them.

#jira UE-148232
#rb brandon.schaefer
#preflight 624c76a8712e7c09a6a15d69

[CL 19627733 by matt johnson in ue5-main branch]